EdTech/cosine/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
transformers
torch
streamlit
requests
google-generativeai
huggingface_hub

huggingface_hub[hf_xet]
